requirements.txt
requests>=2.31
pdfplumber>=0.11
fastapi>=0.115
uvicorn>=0.34
gunicorn>=23.0
pydantic-settings>=2.9
supabase>=2.15
openai>=1.30
httpx>=0.28
jinja2>=3.1
mcp>=1.0.0
starlette>=0.40
python-dotenv>=1.0
